+++ crates/stackable-operator/src/lib.rs
@@ -1,3 +1,11 @@
+//! ## Crate Features
+//!
+//! - `default` enables a default set of features which most operators need.
+//! - `full` enables all available features.
+//! - `time` enables interoperability between [`time::Duration`] and the `time` crate.
+//! - `telemetry` enables various helpers for emitting telemetry data.
+//! - `versioned` enables the macro for CRD versioning.
+
 pub mod builder;
 pub mod cli;
 pub mod client;
@@ -22,14 +30,16 @@ pub mod time;
 pub mod utils;
 pub mod validation;
 
-// Internal re-exports
-pub use stackable_shared::{crd::CustomResourceExt, yaml::YamlSchema};
-
-pub mod shared {
-    pub use stackable_shared::*;
-}
-
 // External re-exports
-pub use ::k8s_openapi;
-pub use ::kube;
-pub use ::schemars;
+pub use k8s_openapi;
+pub use kube;
+pub use schemars;
+// Internal re-exports
+// TODO (@Techassi): Ideally we would want webhook and certs exported here as
+// well, but that would require some restructuring of crates.
+pub use stackable_shared as shared;
+pub use stackable_shared::{crd::CustomResourceExt, yaml::YamlSchema};
+#[cfg(feature = "telemetry")]
+pub use stackable_telemetry as telemetry;
+#[cfg(feature = "versioned")]
+pub use stackable_versioned as versioned;